refactor(#26): 반의어 확인 버튼 컴포넌트로 분리
