+++ features/switch-opposite-word/index.ts
@@ -1,0 +1,3 @@
+export { useSwitchOpposite } from './model/useSwitchOpposite'
+
+export { default as OppositeWordButton } from './ui/OppositeWordButton'
